+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Added
 
+- Transducer: `sexp` for interpretting a char stream as separate SEXP strings.
 - Reducer: `quantities` for counting unique occurrences of streamed items.
 - Reducer: `partition` for splitting the stream results.